Remove the ReceiverProcessor type and merge its ProcessEvents method into EventReceiver also split EventReceiver into it's own file.
